Move decode results to a standalone header.
